+++ src/components/common/unlock-button/unlock-button.tsx
@@ -1,6 +1,12 @@
 import { Component, h, Prop } from '@stencil/core';
+import classNames from 'classnames';
 import { getIsExtensionAvailable, getIsMetaMaskAvailable } from 'components/visual/unlock-panel/helpers';
+import type { IProviderBase } from 'types/provider.types';
 import { ProviderTypeEnum } from 'types/provider.types';
+
+const unlockButtonClasses: Record<string, string> = {
+  statusIcon: 'mvx:fill-teal-400!',
+};
 
 @Component({
   tag: 'mvx-unlock-button',
@@ -9,8 +15,10 @@ import { ProviderTypeEnum } from 'types/provider.types';
 })
 export class UnlockButton {
   @Prop() label: string;
-  @Prop() icon: HTMLElement;
-  @Prop() type?: ProviderTypeEnum;
+  @Prop() iconUrl: string;
+  @Prop() icon?: HTMLElement;
+
+  @Prop() type?: IProviderBase['type'];
   @Prop() class?: string;
 
   render() {
@@ -23,25 +31,27 @@ export class UnlockButton {
 
     return (
       <div class={{ 'unlock-button': true, [this.class]: Boolean(this.class) }}>
-        {this.icon ? (
-          <div class="unlock-button-icon">{this.icon}</div>
-        ) : (
-          <div class="unlock-button-icon">
-            <mvx-multiversx-logo-icon />
-          </div>
-        )}
-
         <div class="unlock-button-label">{this.label}</div>
+
+        <div
+          class={{
+            'unlock-button-icon': true,
+            'clipped': isExtensionProvider,
+          }}
+        >
+          {this.icon ? this.icon : <img src={this.iconUrl} alt={this.label} />}
+        </div>
 
         {isDetectableProvider && (
           <div class="unlock-button-status">
-            {shouldShowOpenLabel ? (
-              <div class="unlock-button-status-open">Open</div>
-            ) : (
-              <div class="unlock-button-status-install">
-                <span class="unlock-button-status-install-label">Install</span>
-                <mvx-arrow-up-right-icon class="unlock-button-status-install-icon" />
-              </div>
+            <div class="unlock-button-status-text">{shouldShowOpenLabel ? 'Open' : 'Install'}</div>
+
+            {!shouldShowOpenLabel && (
+              <mvx-arrow-up-right-icon
+                class={classNames('unlock-button-status-icon', {
+                  [unlockButtonClasses.statusIcon]: true,
+                })}
+              />
             )}
           </div>
         )}