+++ src/components/visual/pagination/components/pagination-ellipsis-form/pagination-ellipsis-form.tsx
@@ -1,6 +1,10 @@
 import type { EventEmitter } from '@stencil/core';
 import { Component, Event, h, Prop, State, Watch } from '@stencil/core';
-import { BigNumber } from 'bignumber.js';
+import classNames from 'classnames';
+
+const paginationEllipsisFormClasses: Record<string, string> = {
+  buttonIcon: 'mvx:p-[10px] mvx:hover:!fill-teal-400',
+};
 
 @Component({
   tag: 'mvx-pagination-ellipsis-form',
@@ -9,14 +13,14 @@ import { BigNumber } from 'bignumber.js';
 })
 export class PaginationEllipsisForm {
   @State() pageValue: string = '';
-  @Event() search: EventEmitter<number>;
+  @Event({ bubbles: false, composed: false }) search: EventEmitter<number>;
 
   @Prop() maxPageToSearchFor: number;
   @Prop() isVisible: boolean = false;
 
   constructor() {
+    this.handleInput = this.handleInput.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
-    this.handleChange = this.handleChange.bind(this);
     this.handleKeyDown = this.handleKeyDown.bind(this);
     this.handleInputReference = this.handleInputReference.bind(this);
   }
@@ -27,19 +31,25 @@ export class PaginationEllipsisForm {
   }
 
   private handleKeyDown(event: KeyboardEvent) {
+    if (event.code === 'Enter') {
+      event.preventDefault();
+      this.handleSubmit(event);
+    }
+
     if (['Equal', 'Minus', 'Period', 'KeyE', 'Comma'].includes(event.code)) {
       event.preventDefault();
       return;
     }
   }
 
-  private handleChange(event: Event) {
+  private handleInput(event: Event) {
     const input = event.target as HTMLInputElement;
-    const valueBigNumber = new BigNumber(input.value);
-    const isBelowMax = valueBigNumber.isLessThanOrEqualTo(this.maxPageToSearchFor);
+    const isBelowMax = parseFloat(input.value) <= this.maxPageToSearchFor;
 
-    if (valueBigNumber.isNaN() || isBelowMax) {
-      this.pageValue = valueBigNumber.toString();
+    if (isBelowMax) {
+      this.pageValue = input.value;
+    } else {
+      input.value = this.pageValue;
     }
   }
 
@@ -61,7 +71,7 @@ export class PaginationEllipsisForm {
 
   render() {
     return (
-      <form class="pagination-ellipsis-form" onSubmit={this.handleSubmit} onClick={(event: MouseEvent) => event.stopPropagation()}>
+      <div class="pagination-ellipsis-form" onClick={(event: MouseEvent) => event.stopPropagation()}>
         <label htmlFor="paginationSearch" class="pagination-ellipsis-form-field-label">
           Page
         </label>
@@ -74,17 +84,18 @@ export class PaginationEllipsisForm {
             id="paginationSearch"
             value={this.pageValue}
             name="paginationSearch"
-            onInput={this.handleChange}
+            onInput={this.handleInput}
+            max={this.maxPageToSearchFor}
             onKeyDown={this.handleKeyDown}
             ref={this.handleInputReference}
             class="pagination-ellipsis-form-field-input"
           />
 
           <div class="pagination-ellipsis-form-button" onClick={this.handleSubmit}>
-            <mvx-magnifying-glass-icon class="pagination-ellipsis-form-button-icon" />
+            <mvx-magnifying-glass-icon class={classNames('pagination-ellipsis-form-button-icon', paginationEllipsisFormClasses.buttonIcon)} />
           </div>
         </div>
-      </form>
+      </div>
     );
   }
 }