+++ src/components/visual/unlock-panel/components/unlock-provider-button/unlock-provider-button.tsx
@@ -1,53 +1,24 @@
-import { Component, Fragment, h, Prop } from '@stencil/core';
-import { ProviderLabelsEnum, ProviderTypeEnum } from 'types/provider.types';
+import { Component, h, Prop } from '@stencil/core';
+import type { IProviderBase } from 'types/provider.types';
+import type { ProviderTypeEnum } from 'types/provider.types';
 
-const providerButtonInfo = {
-  [ProviderTypeEnum.extension]: {
-    icon: <mvx-extension-provider-icon />,
-    label: ProviderLabelsEnum.extension,
-  },
-  [ProviderTypeEnum.metamask]: {
-    icon: <mvx-metamask-provider-icon />,
-    label: ProviderLabelsEnum.metamask,
-  },
-  [ProviderTypeEnum.passkey]: {
-    icon: <mvx-passkey-provider-icon />,
-    label: ProviderLabelsEnum.passkey,
-  },
-  [ProviderTypeEnum.walletConnect]: {
-    icon: <mvx-multiversx-logo-icon />,
-    label: ProviderLabelsEnum.walletConnect,
-  },
-  [ProviderTypeEnum.ledger]: {
-    icon: <mvx-ledger-provider-icon />,
-    label: ProviderLabelsEnum.ledger,
-  },
-  [ProviderTypeEnum.crossWindow]: {
-    icon: <mvx-wallet-provider-icon />,
-    label: ProviderLabelsEnum.crossWindow,
-  },
-  [ProviderTypeEnum.xalias]: {
-    icon: <mvx-xalias-provider-icon />,
-    label: ProviderLabelsEnum.xalias,
-  },
-};
+import { getProviderButtonIcon } from '../../helpers';
 
 @Component({
   tag: 'mvx-unlock-provider-button',
   shadow: true,
 })
 export class UnlockProviderButton {
-  @Prop() type: ProviderTypeEnum;
-  @Prop() providers: string[];
+  @Prop() provider: IProviderBase<ProviderTypeEnum>;
   @Prop() class?: string;
 
   render() {
-    const walletInfo = this.type ? providerButtonInfo[this.type] : null;
-
-    if (!walletInfo) {
-      return <Fragment />;
+    if (!this.provider) {
+      return null;
     }
 
-    return <mvx-unlock-button icon={walletInfo.icon} label={walletInfo.label} type={this.type} class={this.class} />;
+    const icon: HTMLElement | null = !this.provider.iconUrl ? getProviderButtonIcon(this.provider.type) : null;
+
+    return <mvx-unlock-button iconUrl={this.provider.iconUrl} label={this.provider.name} type={this.provider.type} class={this.class} icon={icon} />;
   }
 }