+++ src/components/visual/unlock-panel/helpers/getProviderButtonIcon.tsx
@@ -1,0 +1,23 @@
+import { h } from '@stencil/core';
+import type { IProviderBase } from 'types/provider.types';
+import { ProviderTypeEnum } from 'types/provider.types';
+
+export const getProviderButtonIcon = (providerType: IProviderBase['type']): HTMLElement => {
+  switch (providerType) {
+    case ProviderTypeEnum.extension:
+      return <mvx-extension-provider-icon />;
+    case ProviderTypeEnum.metamask:
+      return <mvx-metamask-provider-icon />;
+    case ProviderTypeEnum.passkey:
+      return <mvx-passkey-provider-icon />;
+    case ProviderTypeEnum.walletConnect:
+      return <mvx-multiversx-logo-icon />;
+    case ProviderTypeEnum.ledger:
+      return <mvx-ledger-provider-icon />;
+    case ProviderTypeEnum.crossWindow:
+      return <mvx-wallet-provider-icon />;
+
+    default:
+      return <mvx-multiversx-logo-icon />;
+  }
+};