+++ src/components/visual/unlock-panel/helpers/index.ts
@@ -1,2 +1,3 @@
 export * from './getIsExtensionAvailable';
 export * from './getIsMetaMaskAvailable';
+export * from './getProviderButtonIcon';