+++ src/components/visual/unlock-panel/unlock-panel.tsx
@@ -1,7 +1,7 @@
-import type { EventEmitter } from '@stencil/core';
-import { Component, Element, Event, h, Method, Prop, State, Watch } from '@stencil/core';
+import { Component, Element, h, Method, Prop, State, Watch } from '@stencil/core';
 import classNames from 'classnames';
-import { ProviderLabelsEnum, ProviderTypeEnum } from 'types/provider.types';
+import type { IProviderBase } from 'types/provider.types';
+import { ProviderTypeEnum } from 'types/provider.types';
 import type { IEventBus } from 'utils/EventBus';
 import { EventBus } from 'utils/EventBus';
 import { processImgSrc } from 'utils/processImgSrc';
@@ -9,7 +9,9 @@ import { processImgSrc } from 'utils/processImgSrc';
 import { getIsExtensionAvailable, getIsMetaMaskAvailable } from './helpers';
 import { UnlockPanelEventsEnum } from './unlock-panel.types';
 
-const DEFAULT_ALLOWED_PROVIDERS = Object.values(ProviderTypeEnum);
+const unlockPanelClasses: Record<string, string> = {
+  footerIcon: 'mvx:w-4! mvx:h-auto!',
+};
 
 @Component({
   tag: 'mvx-unlock-panel',
@@ -21,18 +23,15 @@ export class UnlockPanel {
   @Element() hostElement: HTMLElement;
 
   @Prop() isOpen: boolean = false;
-  @Prop() allowedProviders?: ProviderTypeEnum[] = DEFAULT_ALLOWED_PROVIDERS;
-
-  @Event() close: EventEmitter;
-  @Event() login: EventEmitter<{ provider: ProviderTypeEnum; anchor?: HTMLElement }>;
+  @Prop() allowedProviders: IProviderBase[] = [];
 
   @State() isLoggingIn: boolean = false;
   @State() isIntroScreenVisible: boolean = false;
-  @State() selectedMethod: ProviderTypeEnum | null = null;
+  @State() selectedMethod: IProviderBase | null = null;
   @State() hasSlotContent: boolean = false;
   @State() panelState = {
     isOpen: false,
-    allowedProviders: DEFAULT_ALLOWED_PROVIDERS,
+    allowedProviders: this.allowedProviders,
   };
   @Method() async getEventBus() {
     return this.eventBus;
@@ -44,15 +43,15 @@ export class UnlockPanel {
   }
 
   @Watch('allowedProviders')
-  handleAllowedProvidersChange(newValue?: ProviderTypeEnum[]) {
-    this.panelState = { ...this.panelState, allowedProviders: newValue ?? DEFAULT_ALLOWED_PROVIDERS };
+  handleAllowedProvidersChange(newValue: IProviderBase[]) {
+    this.panelState = { ...this.panelState, allowedProviders: newValue };
   }
 
-  private isExtensionInstalled(currentProvider: ProviderTypeEnum) {
+  private isExtensionInstalled(currentProvider: IProviderBase['type']) {
     return currentProvider === ProviderTypeEnum.extension && getIsExtensionAvailable();
   }
 
-  private isMetaMaskInstalled(currentProvider: ProviderTypeEnum) {
+  private isMetaMaskInstalled(currentProvider: IProviderBase['type']) {
     return currentProvider === ProviderTypeEnum.metamask && getIsMetaMaskAvailable();
   }
 
@@ -66,47 +65,49 @@ export class UnlockPanel {
     }
 
     this.eventBus.unsubscribe(UnlockPanelEventsEnum.OPEN, this.unlockPanelUpdate.bind(this));
+    this.eventBus.unsubscribe(UnlockPanelEventsEnum.CANCEL_IN_PROVIDER, this.handleResetLoginState.bind(this));
     this.isLoggingIn = false;
     this.selectedMethod = null;
-    this.panelState = { isOpen: false, allowedProviders: DEFAULT_ALLOWED_PROVIDERS };
+    this.panelState = { isOpen: false, allowedProviders: [] };
 
-    return new Promise(resolve => setTimeout(resolve, 300));
+    const delayClosingAnimation = new Promise(resolve => setTimeout(resolve, 300));
+    return delayClosingAnimation;
   }
 
-  private observeContainer(element: HTMLElement | null) {
-    if (!element) {
-      return;
-    }
-
-    this.anchor = element;
-
-    if (this.observer) {
-      this.observer.disconnect();
-    }
-
-    this.observer = new MutationObserver(() => {
-      this.isLoggingIn = element.childElementCount > 0;
-    });
-
-    this.observer.observe(element, { childList: true, subtree: true });
+  componentDidLoad() {
+    this.eventBus.subscribe(UnlockPanelEventsEnum.OPEN, this.unlockPanelUpdate.bind(this));
+    this.eventBus.subscribe(UnlockPanelEventsEnum.CANCEL_IN_PROVIDER, this.handleResetLoginState.bind(this));
   }
 
-  handleLogin(provider: ProviderTypeEnum) {
-    this.eventBus.publish(UnlockPanelEventsEnum.LOGIN, { type: provider, anchor: this.anchor });
+  private unlockPanelUpdate(payload: { isOpen: boolean; allowedProviders: IProviderBase[] }) {
+    this.panelState = {
+      ...payload,
+      allowedProviders: payload.allowedProviders,
+    };
+  }
+
+  handleLogin(provider: IProviderBase) {
     this.selectedMethod = provider;
 
-    switch (provider) {
+    switch (provider.type) {
       case ProviderTypeEnum.ledger:
         this.isIntroScreenVisible = true;
+        break;
+      case ProviderTypeEnum.crossWindow:
+      case ProviderTypeEnum.extension:
+      case ProviderTypeEnum.metamask:
+      case ProviderTypeEnum.passkey:
+        this.isIntroScreenVisible = true;
+        this.isLoggingIn = true;
+        this.eventBus.publish(UnlockPanelEventsEnum.LOGIN, { type: provider.type, anchor: this.anchor });
         break;
       default:
         this.handleAccess();
     }
   }
 
-  handleResetLoginState(event: MouseEvent) {
-    event.preventDefault();
-
+  handleResetLoginState(event?: MouseEvent) {
+    event?.preventDefault?.();
     this.isLoggingIn = false;
     this.isIntroScreenVisible = false;
     this.selectedMethod = null;
@@ -133,17 +134,19 @@ export class UnlockPanel {
   handleAccess() {
     this.isIntroScreenVisible = false;
     this.isLoggingIn = true;
-    this.login.emit({ provider: this.selectedMethod, anchor: this.anchor });
+    this.eventBus.publish(UnlockPanelEventsEnum.LOGIN, { type: this.selectedMethod.type, anchor: this.anchor });
   }
 
   render() {
-    const detectedProviders: ProviderTypeEnum[] = this.panelState.allowedProviders.filter(
-      allowedProvider => this.isExtensionInstalled(allowedProvider) || this.isMetaMaskInstalled(allowedProvider),
+    const detectedProviders: IProviderBase[] = this.panelState.allowedProviders.filter(
+      allowedProvider => this.isExtensionInstalled(allowedProvider.type) || this.isMetaMaskInstalled(allowedProvider.type),
     );
 
     const otherProviders = this.panelState.allowedProviders.filter(allowedProvider => !detectedProviders.includes(allowedProvider));
-    const panelTitle = this.selectedMethod ? ProviderLabelsEnum[this.selectedMethod] : 'Connect your wallet';
+    const panelTitle = this.selectedMethod ? this.selectedMethod.name : 'Connect your wallet';
     const hasDetectedProviders = detectedProviders.length > 0;
+
+    const isProviderScreenVisible = !this.isLoggingIn && !this.isIntroScreenVisible;
 
     return (
       <mvx-side-panel
@@ -153,10 +156,17 @@ export class UnlockPanel {
         onClose={this.handleClose.bind(this)}
         onBack={this.handleResetLoginState.bind(this)}
       >
-        <div id="anchor" ref={element => this.observeContainer(element)} class={{ 'unlock-panel-anchor': this.isLoggingIn }} />
-        {this.isIntroScreenVisible && <mvx-unlock-provider-intro provider={this.selectedMethod} onAccess={this.handleAccess.bind(this)} />}
+        <div
+          id="anchor"
+          ref={el => {
+            this.anchor = el;
+          }}
+          class={{ 'unlock-panel-anchor': this.isLoggingIn }}
+        >
+          {this.isIntroScreenVisible && <mvx-unlock-provider-intro provider={this.selectedMethod} onAccess={this.handleAccess.bind(this)} />}
+        </div>
 
-        {!this.isLoggingIn && !this.isIntroScreenVisible && (
+        {isProviderScreenVisible && (
           <div class="unlock-panel">
             <div class="unlock-panel-groups">
               {hasDetectedProviders && (
@@ -164,15 +174,8 @@ export class UnlockPanel {
                   <div class="unlock-panel-group-label">Detected</div>
 
                   <div class="unlock-panel-group-providers">
-                    {detectedProviders.map((provider, providerIndex) => (
-                      <mvx-unlock-provider-button
-                        type={provider}
-                        onClick={this.handleLogin.bind(this, provider)}
-                        class={classNames('unlock-panel-group-provider', {
-                          first: providerIndex === 0,
-                          last: providerIndex === detectedProviders.length - 1,
-                        })}
-                      />
+                    {detectedProviders.map(provider => (
+                      <mvx-unlock-provider-button provider={provider} onClick={this.handleLogin.bind(this, provider)} />
                     ))}
                   </div>
                 </div>
@@ -182,15 +185,8 @@ export class UnlockPanel {
                 <div class="unlock-panel-group-label">{hasDetectedProviders ? 'Other Options' : 'Options'}</div>
 
                 <div class="unlock-panel-group-providers">
-                  {otherProviders.map((provider, providerIndex) => (
-                    <mvx-unlock-provider-button
-                      type={provider}
-                      onClick={this.handleLogin.bind(this, provider)}
-                      class={classNames('unlock-panel-group-provider', {
-                        first: providerIndex === 0,
-                        last: providerIndex === otherProviders.length - 1,
-                      })}
-                    />
+                  {otherProviders.map(provider => (
+                    <mvx-unlock-provider-button provider={provider} onClick={this.handleLogin.bind(this, provider)} />
                   ))}
 
                   <slot />
@@ -207,23 +203,16 @@ export class UnlockPanel {
                   Take full control of <br /> your assets.
                 </div>
 
-                <mvx-arrow-up-right-icon class="unlock-panel-footer-icon" />
+                <mvx-arrow-up-right-icon
+                  class={classNames('unlock-panel-footer-icon', {
+                    [unlockPanelClasses.footerIcon]: true,
+                  })}
+                />
               </div>
             </div>
           </div>
         )}
       </mvx-side-panel>
     );
-  }
-
-  componentDidLoad() {
-    this.eventBus.subscribe(UnlockPanelEventsEnum.OPEN, this.unlockPanelUpdate.bind(this));
-  }
-
-  private unlockPanelUpdate(payload: { isOpen: boolean; allowedProviders: ProviderTypeEnum[] }) {
-    this.panelState = {
-      ...payload,
-      allowedProviders: payload.allowedProviders ?? DEFAULT_ALLOWED_PROVIDERS,
-    };
   }
 }